First Commit
This is a repository that I made to complete the first project of CS50's Web Programming with Python and JavaScript. My work is done, although this is my first commit in Github, as I prepared my work days before I watched Week 1 video on Git and learned how to use GitHub.